Move PyResult from splash.utils to splash.lua
